README improvements (#7)
* docs(readme): Prometheus with capital letter

It's a name.

* docs(readme): correct "Capability" header

It lists multiple capabilities, so should be in plural.

* docs(readme): prefix all shell commands with `$`

All other commands are prefixed that way. It was missing in two places.

* docs(readme): bullet indentation and textwrapping

Nicer to the eye.

* docs(readme): colon post enumeration

* docs(readme): refer to tests for adding newer Kafka support

* Revert "docs(readme): refer to tests for adding newer Kafka support"

Code will change and be refactored. In fact #9 already changes the
location of the script. Dropping this piece of documentation. See also
[1].

[1] https://github.com/kawamuray/prometheus-kafka-consumer-group-exporter/pull/7#discussion_r104138100

This reverts commit 645a01f8ae3d326745d8f828069d1977e8e71a30.